Improve task exit status detection.
We had to restore default SIGCHLD handler in the MiscCommon::execute wrapper. If we fail to do so, we might fail to waitpid our children.
After we started using boost::process we noticed that ::waitpid fails.
boost:process either sets its own handler or there is a call for signal(SIGCHLD, SIG_IGN)