Implement SM write queue draining
- Instead of dropping messages by timeout, we implement a write queue draining logic.
For the moment it is used only for SM Intercom channels. when a user task exists, the corresponding Intercom channel is automatically drained from pending messages.
This prevents blocking transport threads and also prevent us from prematurely dropping messages which could happen if a user task is too slow to start.